refactor(architecture): Migrate to Centralized Auth & Standardize ID Types
- **Global Architecture Changes:**
  - Migrated from Distributed Auth to **Centralized Authentication**.
  - Standardized `userId` type from `Integer` to `Long` across all microservices for scalability.

- **Identity Service:**
  - Exposed `/validate` endpoint to verify tokens and return `userId` (Long).
  - Updated `User` entity, Repository, and `JwtService` to support `Long` IDs.

- **Card & Payment Services:**
  - **Security:** Removed local `JwtService` and `secret-key` configs (Keys are now isolated in Identity Service).
  - **Integration:** Implemented `IdentityServiceClient` (Feign) to delegate token validation.
  - **Refactor:** Updated `JwtAuthenticationFilter` to use the new Feign client.
  - **Cleanup:** Deleted redundant `IdentityAdapter`, `IdentityClient`, and duplicate user existence checks.

+++ identity-service/src/main/java/com/hk/hk/identityservice/controller/UserController.java
@@ -12,8 +12,7 @@ public class UserController {
     private final AuthService authService;
 
     @GetMapping("/{id}/exists")
-    public boolean existsById(@PathVariable Integer id) {
-        // Az önce AuthService'e eklediğimiz metodu çağırıyoruz
+    public boolean existsById(@PathVariable Long id) {
         return authService.existsById(id);
     }
 }
